generalized a bunch of code
