+++ backend/db-helpers.js
@@ -1,0 +1,39 @@
+/*Basically a bunch of functions that makes db function into promises*/
+const sqlite3 = require("sqlite3");
+
+/**
+ * 
+ * @param {sqlite3.Database} db the database to target
+ * @param {string} query the sql query to execute
+ * @param {array<string>} param the parameters to replace ? with
+ * @returns {Promise<Array<object>>} the rows it found
+ */
+exports.db_all = async function db_all(db, query, param) {
+	return await new Promise((resolve, reject) => {
+		db.all(query, param, (err, rows) => {
+			if (err) {
+				reject(err);
+			} else {
+			 	resolve(rows);
+			}
+		});
+	}) 
+}
+
+/**
+ * @param {*} db the database to target
+ * @param {*} query the sql query to execute
+ * @param {*} param the parameters to replace ? with
+ * @returns {Promise<object>} the row it found
+ */
+exports.db_get = async function db_get(db, query, param) {
+	return await new Promise((resolve, reject) => {
+		db.get(query, param, (err, row) => {
+			if (err) {
+				reject(err);
+			} else {
+				resolve(row);
+			}
+		})
+	})
+}
